==> 04-Individual-Populations/10_pca_by_population.Rmd <==
```{r echo=FALSE,  message=FALSE, warning=FALSE}
library(tidyverse)
library(gridExtra)
library(ggsignif)
library(ggpointdensity)
library(ggpubr)
library(ggrepel)
```

```{r echo=FALSE, message=FALSE, warning=FALSE}
# Load data, rename and merge the populations in a single data frame

pop1_genes <- read.table('input/2023-08-04-ranked_by_dens-POP-1.txt', 
                         header = TRUE)
pop1_genes <- pop1_genes[, c("gene", "clr", "ncd05")]
names(pop1_genes) <- c("gene", "clr_pop1", "ncd05_pop1")

pop2_genes <- read.table('input/2023-08-04-ranked_by_dens-POP-2.txt', 
                         header = TRUE)
pop2_genes <- pop2_genes[, c("gene", "clr", "ncd05")]
names(pop2_genes) <- c("gene", "clr_pop2", "ncd05_pop2")

pop3_genes <- read.table('input/2023-08-04-ranked_by_dens-POP-3.txt', 
                         header = TRUE)
pop3_genes <- pop3_genes[, c("gene", "clr", "ncd05")]
names(pop3_genes) <- c("gene", "clr_pop3", "ncd05_pop3")

pop4_genes <- read.table('input/2023-08-04-ranked_by_dens-POP-4.txt', 
                         header = TRUE)
pop4_genes <- pop4_genes[, c("gene", "clr", "ncd05")]
names(pop4_genes) <- c("gene", "clr_pop4", "ncd05_pop4")

merged_pops <- merge(
  merge(
    merge(
      pop3_genes, pop2_genes, by = "gene"), 
    pop1_genes, by = "gene"), 
  pop4_genes, by = "gene"
  )

```

Head of input file to be standardized and then used in prcomp

```{r echo=TRUE, message=FALSE, warning=FALSE}

head(merged_pops)
#Standardize data 

gene_names <- merged_pops$gene
standardized_data <- scale(merged_pops[, -1])

# Perform PCA
pca_result <- prcomp(standardized_data, center = TRUE, scale. = TRUE)

# Create a data frame with PCA scores
pca_scores <- as.data.frame(pca_result$x)
#colnames(pca_scores) <- c("PC1", "PC2")  # Rename columns
```

Explained variance 

```{r echo=TRUE, message=FALSE, warning=FALSE}

explained_variance <- (pca_result$sdev^2) / sum(pca_result$sdev^2)
explained_variance

```


```{r echo=FALSE, message=FALSE, warning=FALSE}

# Calculate cumulative explained variance
cumulative_variance <- cumsum(pca_result$sdev^2) / sum(pca_result$sdev^2)
cumulative_variance

# Create a scree plot
scree_data <- data.frame(
  PC = 1:length(pca_result$sdev),
  Cumulative_Variance = cumulative_variance
)

ggplot(scree_data, aes(x = PC, y = Cumulative_Variance)) +
  geom_line() +
  geom_point() +
  labs(x = "Number of Principal Components", y = "Cumulative Explained Variance") +
  ggtitle("Scree Plot: Cumulative Explained Variance") +
  theme_minimal()

```



```{r echo=FALSE, message=FALSE, warning=FALSE}
# Add gene names back
pca_results_with_names <- data.frame(
  gene = gene_names,
  PC1 = pca_scores$PC1,
  PC2 = pca_scores$PC2,
  PC3 = pca_scores$PC3,
  PC4 = pca_scores$PC4,
  PC5 = pca_scores$PC5,
  PC6 = pca_scores$PC6,
  PC7 = pca_scores$PC7,
  PC8 = pca_scores$PC8
)

#get the master table
bias_and_all <- read.table('input/2023_08_22_MASTER_cleaned_all_pops.txt',
                             header = TRUE) 


bias_and_all <- merge(bias_and_all, pca_results_with_names, by = "gene")
bias_and_all$candens <- as.factor(bias_and_all$candens)
bias_and_all$hymenop <- as.factor(bias_and_all$hymenop)
bias_and_all$pop1 <- as.factor(bias_and_all$pop1)
bias_and_all$zeros <- "0"

```

```{r echo=TRUE, message=FALSE, warning=FALSE}

strong_genes <- subset(bias_and_all, PC1 >= 5) %>% arrange(desc(PC1))

write.table(strong_genes, file = "stong_genes_pca.txt", quote = FALSE,
            row.names = FALSE, col.names = FALSE)


ensemble_names <- read.table('pca_genes_ensembl.txt', 
                         header = TRUE, sep = "\t")

strong_genes <- merge(strong_genes, ensemble_names, by = "gene" )

#
gene1 <- subset(strong_genes, gene == "LOC105207562")
gene2 <- subset(strong_genes, gene == "LOC105200045")
gene3 <- subset(strong_genes, gene == "LOC105202344")

three_genes <- rbind(gene1, gene2, gene3)

```



```{r echo=TRUE, message=FALSE, warning=FALSE}

ggplot(bias_and_all %>% arrange(candens), 
       aes(x = PC1, y = PC2, colour = candens)) +
  geom_point(size = 4, alpha = 0.6) +
  scale_colour_viridis_d(option = "D", direction = 1, name = "BLS",
                         labels = c("No","Yes")) +
  labs(subtitle = "Candidates from previous analysis") +
  #guides(shape = FALSE) +
  theme_minimal()

```

```{r echo=TRUE, message=FALSE, warning=FALSE}

#quantile(bias_and_all$PC1, 0.01)

ggplot(bias_and_all %>% arrange(candens), 
       aes(x = PC1, y = zeros, colour = candens)) +
  geom_point(size = 4, alpha = 0.6) +
  scale_colour_viridis_d(option = "D", direction = -1, name = "BLS",
                         labels = c("No","Yes")) +
  geom_label_repel(data = three_genes, aes(label = name), colour = "black",
                   size = 2, nudge_x = 2, nudge_y = 0.45, direction = "both",
                   box.padding = 0.3, max.overlaps = 50) +
  labs(title = "Only PC1", subtitle = "Candidates from previous analysis") +
  geom_boxplot(aes(x = PC1, y = zeros), width = 0.2, 
               position = position_dodge(width = 0.5))+
  #guides(shape = FALSE) +
  theme_minimal()

```
```{r echo=TRUE, message=FALSE, warning=FALSE}

#quantile(bias_and_all$PC1, 0.01)

ggplot(bias_and_all %>% arrange(candens), 
       aes(x = PC1, y = zeros)) +
  geom_point(size = 4, alpha = 0.6, colour = "#6CC78A") +
  geom_label_repel(data = three_genes, aes(label = name), colour = "black",
                   size = 2, nudge_x = 2, nudge_y = 0.45, direction = "both",
                   box.padding = 0.3, max.overlaps = 50) +
  labs(subtitle = "Only PC1")+
  geom_boxplot(aes(x = PC1, y = -0.1), width = 0.2)+
  #guides(shape = FALSE) +
  theme_minimal()

```


```{r echo=TRUE, message=FALSE, warning=FALSE}

bls_yes <- subset(bias_and_all, candens == "1")
bls_no <- subset(bias_and_all, candens == "0")

ks.test(bls_no$PC1, bls_yes$PC1)
#ks.test(bls_no$PC1, bls_yes$PC1, alternative = "greater")
ks.test(bls_yes$PC1, bls_no$PC1, alternative = "greater")

```

```{r echo=FALSE, message=FALSE, warning=FALSE}
#adjust = 50,
ggplot(bias_and_all %>% arrange(candens), 
       aes(x = PC1, y = PC2)) +
  geom_pointdensity(size = 4, shape = 18) +
  scale_colour_viridis_c(option = "D", direction = 1, name = "Density") + 
  geom_label_repel(data = three_genes, aes(label = name), colour = "black",
                   size = 2, nudge_x = 3, nudge_y = 6, direction = "both",
                   box.padding = 0.3, max.overlaps = 50) +
  labs(x = "PC1 36%", y = "PC2 15%" ) +
  theme_minimal()

```




```{r echo=FALSE, message=FALSE, warning=FALSE}

ggplot(bias_and_all %>% arrange(immune_new), 
       aes(x = PC1, y = PC2, colour = immune_new, shape = candens)) +
  geom_point(size = 4, alpha = 0.6) +
  scale_colour_viridis_c(option = "D", direction = 1) +
  guides(shape = FALSE)+
  theme_minimal()

```






```{r echo=TRUE, message=FALSE, warning=FALSE}

ggplot(bias_and_all %>% arrange(candens)
       , aes(x = PC1, y = PC2, colour = candens, shape = candens)) +
  geom_point(size = 4, alpha = 0.6) +
  geom_label_repel(data = strong_genes, aes(label = gene), colour = "black",
                   size= 2, nudge_x = 3, nudge_y = 6, direction = "both",
                   box.padding = 0.3, max.overlaps = 50) +
  scale_colour_viridis_d(option = "D", direction = -1, name = "Genes in top 1%",
                         labels = c("No","Yes")) + 
  labs(x = "PC1", y = "PC2") + 
  guides(shape = FALSE) +
  theme_minimal()

```


```{r echo=FALSE, message=FALSE, warning=FALSE}

ggplot(bias_and_all %>% arrange(desc(chr))
       , aes(x = PC1, y = PC2)) +
  geom_point(size = 4, alpha = 0.6, colour = "#3F5743") +
  geom_label_repel(data = three_genes, aes(label = name), colour = "black",
                   size = 2, nudge_x = 3, nudge_y = 6, direction = "both",
                   box.padding = 0.3, max.overlaps = 50) +
  theme_minimal()

```




Other PCs

```{r echo=FALSE, message=FALSE, warning=FALSE}

# ggplot(bias_and_all %>% arrange(desc(density)), 
#        aes(x = PC2, y = PC3, colour = density, shape = candens)) +
#   geom_point(size = 4, alpha = 0.6) +
#   scale_colour_viridis_c(option = "D", direction = 1, name = "Density") +
#   labs(subtitle = "NCD and CLR-based density used to select candidates") +
#   guides(shape = FALSE)+
#   theme_minimal()

```
 
```{r echo=FALSE, message=FALSE, warning=FALSE}

# ggplot(bias_and_all %>% arrange(desc(density)), 
#        aes(x = PC3, y = PC4, colour = density, shape = candens)) +
#   geom_point(size = 4, alpha = 0.6) +
#   scale_colour_viridis_c(option = "D", direction = 1, name = "Density") +
#   labs(subtitle = "NCD and CLR-based density used to select candidates") +
#   guides(shape = FALSE)+
#   theme_minimal()

```
